Backend/requirements.txt
google-cloud-pubsub==2.5.0
google-cloud-storage==1.38.0
google-cloud-vision==2.3.1
SQLAlchemy==1.3.12
PyMySQL==0.9.3
requests
Wand==0.6.7
wikipedia